TREV-433 fixes per review

+++ service/src/main/kotlin/fi/espoo/evaka/EvakaEnv.kt
@@ -32,7 +32,7 @@ data class EvakaEnv(
     val httpClientCertificateCheck: Boolean,
     val fiveYearsOldDaycareEnabled: Boolean,
     val mockClock: Boolean,
-    val nrOfDaysFeeDecisionCanBeSentInAdvance: Int
+    val nrOfDaysFeeDecisionCanBeSentInAdvance: Long
 ) {
     companion object {
         fun fromEnvironment(env: Environment): EvakaEnv {
